pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opengbsa"
version = "0.0.1"
description = "OpenGBSA - A comprehensive MM/GBSA analysis package with advanced features"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "H. Ibrahim Özdemir", email = "halil.ibrahim.oozdemir@gmail.com"}
]
maintainers = [
    {name = "H. Ibrahim Özdemir", email = "halil.ibrahim.oozdemir@gmail.com"}
]
keywords = [
    "molecular dynamics",
    "MM/GBSA", 
    "GBSA",
    "binding free energy",
    "drug design",
    "computational chemistry",
    "openmm",
    "mdtraj",
    "openff",
    "opengbsa"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Chemistry",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "openmm>=8.0.0",
    "mdtraj>=1.9.0",
    "openff-toolkit>=0.14.0",
    "openmmforcefields",
    "prolif>=2.0.0",
    "rdkit-pypi>=2022.9.1",
    "MDAnalysis>=2.0.0",
    "h5py>=3.7.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "pyyaml>=6.0",
    "pathlib2>=2.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.10.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
    "mypy>=0.800",
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0"
]
gpu = [
    "cudatoolkit>=11.0"
]
full = [
    "openff-forcefields",
    "nglview>=3.0.0",
    "mbuild>=0.11.0",
    "foyer>=0.11.0"
]

[project.scripts]
opengbsa = "mmgbsa.cli:main"
mmgbsa = "mmgbsa.cli:main"  # Backward compatibility
mmgbsa-runner = "mmgbsa.cli:main"
mmgbsa-cli = "mmgbsa_cli:main"

[project.urls]
Homepage = "https://github.com/halilibrahimozdemir/opengbsa"
Documentation = "https://opengbsa.readthedocs.io/"
Repository = "https://github.com/halilibrahimozdemir/opengbsa"
"Bug Tracker" = "https://github.com/halilibrahimozdemir/opengbsa/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.txt", "*.md"]

[tool.setuptools.packages.find]
where = ["."]
include = ["mmgbsa*"]
exclude = ["test*", "tests*", "examples*"]

# Development tools configuration
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
] 